cloud_server/requirements.txt
# Autonova RMM - Cloud Server Dependencies

fastapi==0.108.0
uvicorn[standard]==0.25.0
python-socketio==5.10.0
redis==5.0.1
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
aiosqlite==0.19.0
pycryptodome==3.19.0
